pyproject.toml
[project]
name = "llama-index-pdf-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "llama-index>=0.14.21",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-llms-openai>=0.7.5",
    "streamlit>=1.56.0",
    "pypdf>=6.10.2",
    "llama-index-readers-file>=0.6.0",
]
